requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.31.0
tqdm>=4.65.0
flask>=2.0.0
flask-wtf>=1.0.0
flask-bootstrap>=3.3.7.1 